@@ -20,15 +20,13 @@
     "type": "git",
     "url": "https://github.com/veo-labs/openveo-rest-nodejs-client.git"
   },
-  "dependencies": {},
   "devDependencies": {
-    "grunt": "0.4.5",
+    "grunt": "1.0.1",
     "grunt-cli": "1.2.0",
     "grunt-contrib-yuidoc": "1.0.0",
-    "grunt-eslint": "18.1.0",
-    "grunt-extend-config": "0.9.3",
-    "grunt-gh-pages": "1.1.0",
-    "pre-commit": "1.1.2",
+    "grunt-eslint": "19.0.0",
+    "grunt-gh-pages": "2.0.0",
+    "pre-commit": "1.2.2",
     "yuidoc-theme-blue": "0.1.9"
   },
   "scripts": {
@@ -39,7 +37,7 @@
     "eslint"
   ],
   "engines": {
-    "node": "4.*",
-    "npm": "2.5.0 - 2.14.7"
+    "node": ">=7.4.0",
+    "npm": ">=4.0.5"
   }
 }
