@@ -1,5 +1,12 @@
 'use strict';
 
+/**
+ * Defines a client to connect to OpenVeo web service.
+ *
+ * @module rest-client
+ * @main rest-client
+ */
+
 const url = require('url');
 const fs = require('fs');
 const path = require('path');
@@ -24,21 +31,14 @@ function rejectAll(requests, error) {
   }
 }
 
-/**
- * REST client for OpenVeo web service.
- *
- * It aims to facilitate the interaction with OpenVeo web service. All authentication aspects are managed by the
- * client and are transparent to the user. Requesting an end point, without being authenticated will automatically
- * authenticate first before calling the end point. If token expired, a new authentication is made automatically.
- *
- * @module rest-client
- * @main rest-client
- * @class OpenVeoClient
- */
 class OpenVeoClient {
 
   /**
    * Creates a client to connect to OpenVeo web service.
+   *
+   * It aims to facilitate the interaction with OpenVeo web service. All authentication aspects are managed by the
+   * client and are transparent to the user. Requesting an end point, without being authenticated, will automatically
+   * authenticate first before calling the end point. If token expired, a new authentication is made automatically.
    *
    * @example
    *
@@ -50,11 +50,12 @@ class OpenVeoClient {
    *                 '/home/test/server.crt'
    *     );
    *
+   * @class OpenVeoClient
    * @constructor
-   * @param {String} webServiceUrl The complete url of the OpenVeo web service (with protocol and port)
-   * @param {String} clientId Application client id
-   * @param {String} clientSecret Application client
-   * @param {String} [certificate] Path to the web service server trusted certificate file
+   * @param {String} webServiceUrl The complete url of the OpenVeo Web Service (with protocol and port)
+   * @param {String} clientId Application's client id
+   * @param {String} clientSecret Application's client
+   * @param {String} [certificate] Path to the Web Service server trusted certificate file
    * @throws {TypeError} Thrown if webServiceUrl, clientId or clientSecret is not a valid String
    */
   constructor(webServiceUrl, clientId, clientSecret, certificate) {
@@ -84,6 +85,7 @@ class OpenVeoClient {
        *
        * @property protocol
        * @type String
+       * @final
        */
       protocol: {value: protocol},
 
@@ -92,6 +94,7 @@ class OpenVeoClient {
        *
        * @property hostname
        * @type String
+       * @final
        */
       hostname: {value: urlChunks.hostname},
 
@@ -100,6 +103,7 @@ class OpenVeoClient {
        *
        * @property port
        * @type Number
+       * @final
        */
       port: {value: port},
 
@@ -108,6 +112,7 @@ class OpenVeoClient {
        *
        * @property url
        * @type String
+       * @final
        */
       url: {value: serverUrl},
 
@@ -116,6 +121,7 @@ class OpenVeoClient {
        *
        * @property clientId
        * @type String
+       * @final
        */
       clientId: {value: clientId},
 
@@ -124,6 +130,7 @@ class OpenVeoClient {
        *
        * @property clientSecret
        * @type String
+       * @final
        */
       clientSecret: {value: clientSecret},
 
@@ -132,8 +139,9 @@ class OpenVeoClient {
        *
        * @property credentials
        * @type String
+       * @final
        */
-      credentials: {value: new Buffer(`${clientId}:${clientSecret}`).toString('base64')},
+      credentials: {value: Buffer.from(`${clientId}:${clientSecret}`).toString('base64')},
 
       /**
        * Application access token provided by the web service.
@@ -141,13 +149,14 @@ class OpenVeoClient {
        * @property accessToken
        * @type String
        */
-      accessToken: {writable: true},
+      accessToken: {value: null, writable: true},
 
       /**
        * Path to the web service server certificate file.
        *
        * @property certificate
        * @type String
+       * @final
        */
       certificate: {value: certificate},
 
@@ -175,6 +184,7 @@ class OpenVeoClient {
      *
      * @property authenticateRequest
      * @type Request
+     * @final
      */
     Object.defineProperty(this, 'authenticateRequest', {
       value: this.buildRequest({
@@ -463,10 +473,13 @@ class OpenVeoClient {
 
     // Add web service certificate as a trusted certificate
     if (this.certificate && this.protocol === 'https') {
+
+      /* eslint no-sync: 0 */
       Object.assign(options, {
         rejectUnauthorized: process.env.NODE_ENV === 'production',
         ca: fs.readFileSync(path.normalize(this.certificate))
       });
+
     }
 
     const request = new Request(this.protocol, options, body);
