@@ -1,19 +1,17 @@
 'use strict';
 
+/**
+ * @module rest-client
+ */
+
 const timers = require('timers');
 
-/**
- * A web service request.
- *
- * @module rest-client
- * @main rest-client
- * @class Request
- */
 class Request {
 
   /**
    * Creates a REST request which can be executed / aborted.
    *
+   * @class Request
    * @constructor
    * @param {String} protocol The protocol to use for the request (either 'http' or 'https')
    * @param {Object} [options] The complete list of http(s) options as described by NodeJS http.request
@@ -38,6 +36,7 @@ class Request {
        *
        * @property protocol
        * @type String
+       * @final
        */
       protocol: {value: protocol},
 
@@ -46,6 +45,7 @@ class Request {
        *
        * @property options
        * @type Object
+       * @final
        */
       options: {value: options},
 
@@ -54,6 +54,7 @@ class Request {
        *
        * @property body
        * @type String
+       * @final
        */
       body: {value: body},
 
